separate roadmap items into separate folder, update data collection with sparse control cycling
